refactor(Core): Refactor the DI story for the runtime/reflective injector.
This allows more sweeping changes and optimizations to what we allow here, and
also allows removing all public fields/members from the Provider class itself,
which is a leaky/unsupported abstraction.

PiperOrigin-RevId: 176030536
