+++ _goldens/test_files/di/generated_injectors.dart
@@ -3,7 +3,7 @@ import 'package:angular/angular.dart';
 import 'generated_injectors.template.dart' as ng;
 
 @Injector.generate(const [
-  const ProviderUseClass<Example, ExamplePrime>(),
+  const Provider(Example, useClass: ExamplePrime),
 ])
 Injector doGenerate() => ng.doGenerate$Injector();
