+++ _goldens/test_files/providers.dart
@@ -30,7 +30,6 @@ import 'package:angular/angular.dart';
       useValueString,
       useValue: 'foo',
     ),
-    const ProviderUseMulti<MyMulti>.ofType(MyMulti),
     const Provider(useValueList, useValue: const [
       const MyUseValue('Andrew'),
       const MyUseValue('Matan'),
