+++ _goldens/test_files/providers.template_debug.golden
@@ -77,7 +77,6 @@ List<StaticNodeDebugInfo> nodeDebugInfos_ProvidersComponentHost0 = [
         import2.MyUseExistingNested,
         import2.MyUseValue,
         const OpaqueToken<dynamic>('useValueString'),
-        import2.MyMulti,
         const OpaqueToken<dynamic>('useValueList'),
         const OpaqueToken<dynamic>('useValueMap'),
         const OpaqueToken<dynamic>('useEnums')
@@ -94,10 +93,9 @@ class _ViewProvidersComponentHost0 extends DebugAppView<dynamic> {
   import2.MyUseClass __MyUseClass_0_7;
   import2.MyUseValue __MyUseValue_0_10;
   String __const_OpaqueToken_dynamic___useValueString___0_11;
-  List<import2.MyMulti> __MyMulti_0_12;
-  List<dynamic> __const_OpaqueToken_dynamic___useValueList___0_13;
-  Map<String, dynamic> __const_OpaqueToken_dynamic___useValueMap___0_14;
-  dynamic __const_OpaqueToken_dynamic___useEnums___0_15;
+  List<dynamic> __const_OpaqueToken_dynamic___useValueList___0_12;
+  Map<String, dynamic> __const_OpaqueToken_dynamic___useValueMap___0_13;
+  dynamic __const_OpaqueToken_dynamic___useEnums___0_14;
   static RenderComponentType _renderType;
   _ViewProvidersComponentHost0(AppView<dynamic> parentView, num parentIndex)
       : super(
@@ -153,43 +151,35 @@ class _ViewProvidersComponentHost0 extends DebugAppView<dynamic> {
     return this.__const_OpaqueToken_dynamic___useValueString___0_11;
   }
 
-  List<import2.MyMulti> get _MyMulti_0_12 {
+  List<dynamic> get _const_OpaqueToken_dynamic___useValueList___0_12 {
     dbg(0, 0, 0);
-    if ((this.__MyMulti_0_12 == null)) {
-      (__MyMulti_0_12 = [null]);
-    }
-    return this.__MyMulti_0_12;
-  }
-
-  List<dynamic> get _const_OpaqueToken_dynamic___useValueList___0_13 {
-    dbg(0, 0, 0);
-    if ((this.__const_OpaqueToken_dynamic___useValueList___0_13 == null)) {
-      (__const_OpaqueToken_dynamic___useValueList___0_13 = const [
+    if ((this.__const_OpaqueToken_dynamic___useValueList___0_12 == null)) {
+      (__const_OpaqueToken_dynamic___useValueList___0_12 = const [
         const import2.MyUseValue('Andrew'),
         const import2.MyUseValue('Matan'),
         const import2.MyUseValue.named(optional: true)
       ]);
     }
-    return this.__const_OpaqueToken_dynamic___useValueList___0_13;
+    return this.__const_OpaqueToken_dynamic___useValueList___0_12;
   }
 
-  Map<String, dynamic> get _const_OpaqueToken_dynamic___useValueMap___0_14 {
+  Map<String, dynamic> get _const_OpaqueToken_dynamic___useValueMap___0_13 {
     dbg(0, 0, 0);
-    if ((this.__const_OpaqueToken_dynamic___useValueMap___0_14 == null)) {
-      (__const_OpaqueToken_dynamic___useValueMap___0_14 = const {
+    if ((this.__const_OpaqueToken_dynamic___useValueMap___0_13 == null)) {
+      (__const_OpaqueToken_dynamic___useValueMap___0_13 = const {
         'Andrew': const import2.MyUseValue('Andrew'),
         'Matan': const import2.MyUseValue('Matan')
       });
     }
-    return this.__const_OpaqueToken_dynamic___useValueMap___0_14;
+    return this.__const_OpaqueToken_dynamic___useValueMap___0_13;
   }
 
-  dynamic get _const_OpaqueToken_dynamic___useEnums___0_15 {
+  dynamic get _const_OpaqueToken_dynamic___useEnums___0_14 {
     dbg(0, 0, 0);
-    if ((this.__const_OpaqueToken_dynamic___useEnums___0_15 == null)) {
-      (__const_OpaqueToken_dynamic___useEnums___0_15 = import2.MyEnum.first);
+    if ((this.__const_OpaqueToken_dynamic___useEnums___0_14 == null)) {
+      (__const_OpaqueToken_dynamic___useEnums___0_14 = import2.MyEnum.first);
     }
-    return this.__const_OpaqueToken_dynamic___useEnums___0_15;
+    return this.__const_OpaqueToken_dynamic___useEnums___0_14;
   }
 
   @override
@@ -230,20 +220,17 @@ class _ViewProvidersComponentHost0 extends DebugAppView<dynamic> {
         (0 == nodeIndex))) {
       return _const_OpaqueToken_dynamic___useValueString___0_11;
     }
-    if ((identical(token, import2.MyMulti) && (0 == nodeIndex))) {
-      return _MyMulti_0_12;
-    }
     if ((identical(token, const OpaqueToken<dynamic>('useValueList')) &&
         (0 == nodeIndex))) {
-      return _const_OpaqueToken_dynamic___useValueList___0_13;
+      return _const_OpaqueToken_dynamic___useValueList___0_12;
     }
     if ((identical(token, const OpaqueToken<dynamic>('useValueMap')) &&
         (0 == nodeIndex))) {
-      return _const_OpaqueToken_dynamic___useValueMap___0_14;
+      return _const_OpaqueToken_dynamic___useValueMap___0_13;
     }
     if ((identical(token, const OpaqueToken<dynamic>('useEnums')) &&
         (0 == nodeIndex))) {
-      return _const_OpaqueToken_dynamic___useEnums___0_15;
+      return _const_OpaqueToken_dynamic___useEnums___0_14;
     }
     return notFoundResult;
   }
