+++ _goldens/test_files/providers.template_release.golden
@@ -64,10 +64,9 @@ class _ViewProvidersComponentHost0 extends AppView<dynamic> {
   import1.MyUseClass __MyUseClass_0_7;
   import1.MyUseValue __MyUseValue_0_10;
   String __const_OpaqueToken_dynamic___useValueString___0_11;
-  List<import1.MyMulti> __MyMulti_0_12;
-  List<dynamic> __const_OpaqueToken_dynamic___useValueList___0_13;
-  Map<String, dynamic> __const_OpaqueToken_dynamic___useValueMap___0_14;
-  dynamic __const_OpaqueToken_dynamic___useEnums___0_15;
+  List<dynamic> __const_OpaqueToken_dynamic___useValueList___0_12;
+  Map<String, dynamic> __const_OpaqueToken_dynamic___useValueMap___0_13;
+  dynamic __const_OpaqueToken_dynamic___useEnums___0_14;
   static RenderComponentType _renderType;
   _ViewProvidersComponentHost0(AppView<dynamic> parentView, num parentIndex)
       : super(import3.ViewType.HOST, {}, parentView, parentIndex,
@@ -113,39 +112,32 @@ class _ViewProvidersComponentHost0 extends AppView<dynamic> {
     return this.__const_OpaqueToken_dynamic___useValueString___0_11;
   }
 
-  List<import1.MyMulti> get _MyMulti_0_12 {
-    if ((this.__MyMulti_0_12 == null)) {
-      (__MyMulti_0_12 = [null]);
-    }
-    return this.__MyMulti_0_12;
-  }
-
-  List<dynamic> get _const_OpaqueToken_dynamic___useValueList___0_13 {
-    if ((this.__const_OpaqueToken_dynamic___useValueList___0_13 == null)) {
-      (__const_OpaqueToken_dynamic___useValueList___0_13 = const [
+  List<dynamic> get _const_OpaqueToken_dynamic___useValueList___0_12 {
+    if ((this.__const_OpaqueToken_dynamic___useValueList___0_12 == null)) {
+      (__const_OpaqueToken_dynamic___useValueList___0_12 = const [
         const import1.MyUseValue('Andrew'),
         const import1.MyUseValue('Matan'),
         const import1.MyUseValue.named(optional: true)
       ]);
     }
-    return this.__const_OpaqueToken_dynamic___useValueList___0_13;
+    return this.__const_OpaqueToken_dynamic___useValueList___0_12;
   }
 
-  Map<String, dynamic> get _const_OpaqueToken_dynamic___useValueMap___0_14 {
-    if ((this.__const_OpaqueToken_dynamic___useValueMap___0_14 == null)) {
-      (__const_OpaqueToken_dynamic___useValueMap___0_14 = const {
+  Map<String, dynamic> get _const_OpaqueToken_dynamic___useValueMap___0_13 {
+    if ((this.__const_OpaqueToken_dynamic___useValueMap___0_13 == null)) {
+      (__const_OpaqueToken_dynamic___useValueMap___0_13 = const {
         'Andrew': const import1.MyUseValue('Andrew'),
         'Matan': const import1.MyUseValue('Matan')
       });
     }
-    return this.__const_OpaqueToken_dynamic___useValueMap___0_14;
+    return this.__const_OpaqueToken_dynamic___useValueMap___0_13;
   }
 
-  dynamic get _const_OpaqueToken_dynamic___useEnums___0_15 {
-    if ((this.__const_OpaqueToken_dynamic___useEnums___0_15 == null)) {
-      (__const_OpaqueToken_dynamic___useEnums___0_15 = import1.MyEnum.first);
+  dynamic get _const_OpaqueToken_dynamic___useEnums___0_14 {
+    if ((this.__const_OpaqueToken_dynamic___useEnums___0_14 == null)) {
+      (__const_OpaqueToken_dynamic___useEnums___0_14 = import1.MyEnum.first);
     }
-    return this.__const_OpaqueToken_dynamic___useEnums___0_15;
+    return this.__const_OpaqueToken_dynamic___useEnums___0_14;
   }
 
   @override
@@ -185,20 +177,17 @@ class _ViewProvidersComponentHost0 extends AppView<dynamic> {
         (0 == nodeIndex))) {
       return _const_OpaqueToken_dynamic___useValueString___0_11;
     }
-    if ((identical(token, import1.MyMulti) && (0 == nodeIndex))) {
-      return _MyMulti_0_12;
-    }
     if ((identical(token, const OpaqueToken<dynamic>('useValueList')) &&
         (0 == nodeIndex))) {
-      return _const_OpaqueToken_dynamic___useValueList___0_13;
+      return _const_OpaqueToken_dynamic___useValueList___0_12;
     }
     if ((identical(token, const OpaqueToken<dynamic>('useValueMap')) &&
         (0 == nodeIndex))) {
-      return _const_OpaqueToken_dynamic___useValueMap___0_14;
+      return _const_OpaqueToken_dynamic___useValueMap___0_13;
     }
     if ((identical(token, const OpaqueToken<dynamic>('useEnums')) &&
         (0 == nodeIndex))) {
-      return _const_OpaqueToken_dynamic___useEnums___0_15;
+      return _const_OpaqueToken_dynamic___useEnums___0_14;
     }
     return notFoundResult;
   }
