+++ _tests/test/public_apis.dart
@@ -29,8 +29,6 @@ const NG_CORE = const [
   'APP_ID',
   'Attribute',
   'Provider',
-  'ProviderUseClass',
-  'ProviderUseMulti',
   'ChangeDetectionStrategy',
   'ChangeDetectorRef',
   'Component',
@@ -156,8 +154,6 @@ const angularApis = const [
   'PipeTransform',
   'PlatformRef',
   'Provider',
-  'ProviderUseClass',
-  'ProviderUseMulti',
   'QueryList',
   'ReflectiveInjector',
   'RenderComponentType',
@@ -201,8 +197,6 @@ const diApis = const [
   'Pipe',
   'PipeTransform',
   'Provider',
-  'ProviderUseClass',
-  'ProviderUseMulti',
   'ReflectiveInjector',
   'Self',
   'SkipSelf',
