+++ angular/lib/src/core/di.dart
@@ -1,8 +1,7 @@
 /// A dependency Injection container.
 export "../di/injector/injector.dart" show Injector;
-export "../di/injector/reflective.dart" show ReflectiveInjector;
+export "../di/injector/runtime.dart" show ReflectiveInjector;
 export "../di/module.dart" hide internalModuleToList;
 export "di/decorators.dart";
 export "di/opaque_token.dart" show OpaqueToken;
-export "di/provider.dart"
-    show ProviderUseMulti, Provider, ProviderUseClass, provide, noValueProvided;
+export "di/provider.dart" show Provider, provide, noValueProvided;
