+++ angular/lib/src/core/di/provider.dart
@@ -1,2 +1,2 @@
 // TODO(matanl): Delete once all internal code references ../di/provider.dart.
-export '../../di/provider.dart';
+export '../../di/providers.dart';
