+++ angular/lib/src/platform/bootstrap.dart
@@ -6,7 +6,7 @@ import 'package:angular/src/bootstrap/modules.dart';
 import 'package:angular/src/core/application_ref.dart';
 import 'package:angular/src/core/linker.dart' show ComponentRef;
 import 'package:angular/src/core/testability/testability.dart';
-import 'package:angular/src/di/injector/reflective.dart';
+import 'package:angular/src/di/injector/runtime.dart';
 import 'package:angular/src/platform/browser_common.dart';
 
 export 'browser/tools/tools.dart' show enableDebugTools, disableDebugTools;
